# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Songs:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /songs:
    get:
      consumes:
      - application/json
      description: Get a list of songs with optional filtering and pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Song name
        in: query
        name: song
        type: string
      - description: Group name
        in: query
        name: group
        type: string
      - description: Song text
        in: query
        name: text
        type: string
      - description: Release date
        in: query
        name: releaseDate
        type: string
      - description: Link
        in: query
        name: link
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Songs'
      summary: Get all songs
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Create a new song
      parameters:
      - description: Song details
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Songs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Create a new song
      tags:
      - songs
  /songs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a song by its ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete a song
      tags:
      - songs
    get:
      consumes:
      - application/json
      description: Get a song by its ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      - default: 0
        description: Verse number
        in: query
        name: vers
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Songs'
      summary: Get a song by ID
      tags:
      - songs
    patch:
      consumes:
      - application/json
      description: Update a song by its ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      - description: Song details
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Songs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update a song
      tags:
      - songs
swagger: "2.0"